script/requirements.txt
requests
beautifulsoup4
pandas
geopy
networkx
matplotlib
contextily
geopandas
folium
googlemaps
python-dotenv